feat: home page
- Page d'accueil basique, permettant de naviguer vers les pages teams et Pokémon.
